项目实战/EasyOCR/requirements.txt
# torch
# torchvision>=0.5
opencv-python-headless
scipy
numpy
Pillow
scikit-image
python-bidi
PyYAML
Shapely
pyclipper
ninja
